categories edit

+++ routes/categories.js
@@ -5,6 +5,8 @@ const express = require('express');
 const router = express.Router();
 const auth = require('../middleware/auth');
 const admin = require('../middleware/admin');
+const { Product } = require('../models/product');
+const { Type } = require('../models/type');
 
 
 router.get('/', async(req, res) => {
@@ -44,6 +46,19 @@ router.put('/:id', validateObjectId, async (req, res) => {
 });
 
 router.delete('/:id', [auth, admin], validateObjectId, async (req, res) => {
+    
+    const product = await Product.find({'category._id': req.params.id});
+    const type    = await Type.find({'category._id': req.params.id});
+
+    console.log(product[0]);
+    if(product[0]) {
+        return res.status(404).send({message: 'Cannot delete this category. Products are registered in this category.'});
+    }
+
+    if(type[0]) {
+        return res.status(404).send({message: 'Cannot delete this category. Types are registered in this category'})
+    }
+    
     const category = await Category.findByIdAndRemove(req.params.id);
 
     if(!category) {
